+++ Common/cpp/SharedItems/ShareableValue.cpp
@@ -1,6 +1,12 @@
-#include "Shareable.h"
+#include "ShareableValue.h"
+#include "SharedParent.h"
 #include "NativeReanimatedModule.h"
 #include "Logger.h"
+
+#include "MutableValue.h"
+#include "MutableValueSetterProxy.h"
+#include "RemoteObject.h"
+#include "FrozenObject.h"
 
 namespace reanimated {
 
@@ -15,59 +21,59 @@ void ShareableValue::adaptCache(jsi::Runtime &rt, const jsi::Value &value) {
 }
 
 void ShareableValue::adapt(jsi::Runtime &rt, const jsi::Value &value, ValueType objectType) {
-  if (objectType == MutableValueType) {
-    type = MutableValueType;
+  if (objectType == ValueType::MutableValueType) {
+    type = ValueType::MutableValueType;
     mutableValue = std::make_shared<MutableValue>(rt, value, module);
   } else if (value.isUndefined()) {
-    type = UndefinedType;
+    type = ValueType::UndefinedType;
   } else if (value.isNull()) {
-    type = NullType;
+    type = ValueType::NullType;
   } else if (value.isBool()) {
-    type = BoolType;
+    type = ValueType::BoolType;
     boolValue = value.getBool();
   } else if (value.isNumber()) {
-    type = NumberType;
+    type = ValueType::NumberType;
     numberValue = value.asNumber();
   } else if (value.isString()) {
-    type = StringType;
+    type = ValueType::StringType;
     stringValue = value.asString(rt).utf8(rt);
   } else if (value.isObject()) {
     auto object = value.asObject(rt);
     if (object.isFunction(rt)) {
       if (object.getProperty(rt, "__worklet").isUndefined()) {
         // not a worklet, we treat this as a host function
-        type = HostFunctionType;
+        type = ValueType::HostFunctionType;
         hostRuntime = &rt;
         hostFunction = std::make_shared<jsi::Function>(object.asFunction(rt));
       } else {
         // a worklet
-        type = WorkletFunctionType;
+        type = ValueType::WorkletFunctionType;
         frozenObject = std::make_shared<FrozenObject>(rt, object, module);
       }
     } else if (object.isArray(rt)) {
-      type = ArrayType;
+      type = ValueType::ArrayType;
       auto array = object.asArray(rt);
       for (size_t i = 0, size = array.size(rt); i < size; i++) {
         frozenArray.push_back(adapt(rt, array.getValueAtIndex(rt, i), module));
       }
     } else if (object.isHostObject<MutableValue>(rt)) {
-      type = MutableValueType;
+      type = ValueType::MutableValueType;
       mutableValue = object.getHostObject<MutableValue>(rt);
       adaptCache(rt, value);
     } else if (object.isHostObject<RemoteObject>(rt)) {
-      type = RemoteObjectType;
+      type = ValueType::RemoteObjectType;
       remoteObject = object.getHostObject<RemoteObject>(rt);
       adaptCache(rt, value);
     } else if (object.isHostObject<FrozenObject>(rt)) {
-      type = ObjectType;
+      type = ValueType::ObjectType;
       frozenObject = object.getHostObject<FrozenObject>(rt);
       adaptCache(rt, value);
-    } else if (objectType == RemoteObjectType) {
-      type = RemoteObjectType;
+    } else if (objectType == ValueType::RemoteObjectType) {
+      type = ValueType::RemoteObjectType;
       remoteObject = std::make_shared<RemoteObject>(rt, object, module);
     } else {
       // create frozen object based on a copy of a given object
-      type = ObjectType;
+      type = ValueType::ObjectType;
       frozenObject = std::make_shared<FrozenObject>(rt, object, module);
     }
   } else {
@@ -103,33 +109,33 @@ jsi::Object ShareableValue::createHost(jsi::Runtime &rt, std::shared_ptr<jsi::Ho
 
 jsi::Value ShareableValue::toJSValue(jsi::Runtime &rt) {
   switch (type) {
-    case UndefinedType:
+    case ValueType::UndefinedType:
       return jsi::Value::undefined();
-    case NullType:
+    case ValueType::NullType:
       return jsi::Value::null();
-    case BoolType:
+    case ValueType::BoolType:
       return jsi::Value(boolValue);
-    case NumberType:
+    case ValueType::NumberType:
       return jsi::Value(numberValue);
-    case StringType:
+    case ValueType::StringType:
       return jsi::Value(rt, jsi::String::createFromAscii(rt, stringValue));
-    case ObjectType:
+    case ValueType::ObjectType:
       return createHost(rt, frozenObject);
-    case ArrayType: {
+    case ValueType::ArrayType: {
       jsi::Array array(rt, frozenArray.size());
       for (size_t i = 0; i < frozenArray.size(); i++) {
         array.setValueAtIndex(rt, i, frozenArray[i]->toJSValue(rt));
       }
       return array;
     }
-    case RemoteObjectType:
+    case ValueType::RemoteObjectType:
       if (module->isUIRuntime(rt)) {
         remoteObject->maybeInitializeOnUIRuntime(rt);
       }
       return createHost(rt, remoteObject);
-    case MutableValueType:
+    case ValueType::MutableValueType:
       return createHost(rt, mutableValue);
-    case HostFunctionType:
+    case ValueType::HostFunctionType:
       if (hostRuntime == &rt) {
         // function is accessed from the same runtime it was crated, we just return same function obj
         return jsi::Value(rt, *hostFunction);
@@ -174,7 +180,7 @@ jsi::Value ShareableValue::toJSValue(jsi::Runtime &rt) {
         };
         return jsi::Function::createFromHostFunction(rt, jsi::PropNameID::forAscii(rt, "hostFunction"), 0, clb);
       }
-    case WorkletFunctionType:
+    case ValueType::WorkletFunctionType:
       auto module = this->module;
       auto frozenObject = this->frozenObject;
       if (module->isUIRuntime(rt)) {
@@ -260,179 +266,6 @@ jsi::Value ShareableValue::toJSValue(jsi::Runtime &rt) {
       }
   }
   throw "convert error";
-}
-
-void MutableValueSetterProxy::set(jsi::Runtime &rt, const jsi::PropNameID &name, const jsi::Value &newValue) {
-  auto propName = name.utf8(rt);
-  if (propName == "value") {
-    // you call `this.value` inside of value setter, we should throw
-  } else if (propName == "_value") {
-    mutableValue->setValue(rt, newValue);
-  } else if (propName == "_animation") {
-    // TODO: assert to allow animation to be set from UI only
-    mutableValue->animation = jsi::Value(rt, newValue);
-  }
-}
-
-jsi::Value MutableValueSetterProxy::get(jsi::Runtime &rt, const jsi::PropNameID &name) {
-  auto propName = name.utf8(rt);
-
-  if (propName == "value") {
-    return mutableValue->getValue(rt);
-  } else if (propName == "_value") {
-    return mutableValue->getValue(rt);
-  } else if (propName == "_animation") {
-    return jsi::Value(rt, mutableValue->animation);
-  }
-
-  return jsi::Value::undefined();
-}
-
-void MutableValue::setValue(jsi::Runtime &rt, const jsi::Value &newValue) {
-  std::lock_guard<std::mutex> lock(readWriteMutex);
-  value = ShareableValue::adapt(rt, newValue, module);
-  module->scheduler->scheduleOnUI([this] {
-    for (auto listener : listeners) {
-      listener.second();
-    }
-  });
-}
-
-jsi::Value MutableValue::getValue(jsi::Runtime &rt) {
-  std::lock_guard<std::mutex> lock(readWriteMutex);
-  return value->getValue(rt);
-}
-
-void MutableValue::set(jsi::Runtime &rt, const jsi::PropNameID &name, const jsi::Value &newValue) {
-  auto propName = name.utf8(rt);
-
-  if (module->isHostRuntime(rt)) {
-    if (propName == "value") {
-      auto shareable = ShareableValue::adapt(rt, newValue, module);
-      module->scheduler->scheduleOnUI([this, shareable] {
-        jsi::Runtime &rt = *this->module->runtime.get();
-        auto setterProxy = jsi::Object::createFromHostObject(rt, std::make_shared<MutableValueSetterProxy>(shared_from_this()));
-        jsi::Value newValue = shareable->getValue(rt);
-        module->valueSetter->getValue(rt)
-          .asObject(rt)
-          .asFunction(rt)
-          .callWithThis(rt, setterProxy, newValue);
-      });
-    }
-    return;
-  }
-
-  // UI thread
-  if (propName == "value") {
-    auto setterProxy = jsi::Object::createFromHostObject(rt, std::make_shared<MutableValueSetterProxy>(shared_from_this()));
-    module->valueSetter->getValue(rt)
-      .asObject(rt)
-      .asFunction(rt)
-      .callWithThis(rt, setterProxy, newValue);
-  } else if (propName == "_animation") {
-    // TODO: assert to allow animation to be set from UI only
-    animation = jsi::Value(rt, newValue);
-  }
-}
-
-jsi::Value MutableValue::get(jsi::Runtime &rt, const jsi::PropNameID &name) {
-  auto propName = name.utf8(rt);
-
-  if (propName == "value") {
-    return getValue(rt);
-  }
-
-  if (module->isUIRuntime(rt)) {
-    // _value and _animation should be accessed from UI only
-    if (propName == "_value") {
-      return getValue(rt);
-    } else if (propName == "_animation") {
-      // TODO: assert to allow animation to be read from UI onlu
-      return jsi::Value(rt, animation);
-    }
-  }
-
-  return jsi::Value::undefined();
-}
-
-MutableValue::MutableValue(jsi::Runtime &rt, const jsi::Value &initial, NativeReanimatedModule *module):
-module(module), value(ShareableValue::adapt(rt, initial, module)) {}
-
-unsigned long int MutableValue::addListener(std::function<void ()> listener) {
-  unsigned long id = listeners.size() + 1;
-  listeners.push_back(std::make_pair(id, listener));
-  return id;
-}
-
-void MutableValue::removeListener(unsigned long listenerId) {
-  listeners.erase(std::remove_if(listeners.begin(), listeners.end(), [=](const std::pair<unsigned long, std::function<void()>>& pair) {
-    return pair.first == listenerId;
-  }), listeners.end());
-}
-
-FrozenObject::FrozenObject(jsi::Runtime &rt, const jsi::Object &object, NativeReanimatedModule *module) {
-  auto propertyNames = object.getPropertyNames(rt);
-  for (size_t i = 0, count = propertyNames.size(rt); i < count; i++) {
-    auto propertyName = propertyNames.getValueAtIndex(rt, i).asString(rt);
-    map[propertyName.utf8(rt)] = ShareableValue::adapt(rt, object.getProperty(rt, propertyName), module);
-  }
-}
-
-std::shared_ptr<jsi::Object> FrozenObject::shallowClone(jsi::Runtime &rt) {
-  std::shared_ptr<jsi::Object> object(new jsi::Object(rt));
-  for (auto prop : map) {
-    object->setProperty(rt, jsi::String::createFromUtf8(rt, prop.first), prop.second->getValue(rt));
-  }
-  return object;
-}
-
-jsi::Value FrozenObject::get(jsi::Runtime &rt, const jsi::PropNameID &name) {
-  auto propName = name.utf8(rt);
-  if (propName == "$$typeof") {
-    return jsi::Value(rt, jsi::String::createFromAscii(rt, "object"));
-  } else if (propName == "Symbol.toStringTag") {
-    return jsi::Value(rt, jsi::String::createFromAscii(rt, "FrozenObject"));
-  }
-  auto found = map.find(propName);
-  return found == map.end() ? jsi::Value::undefined() : found->second->getValue(rt);
-}
-
-std::vector<jsi::PropNameID> FrozenObject::getPropertyNames(jsi::Runtime &rt) {
-  std::vector<jsi::PropNameID> result;
-  for (auto it = map.begin(); it != map.end(); it++) {
-    result.push_back(jsi::PropNameID::forUtf8(rt, it->first));
-  }
-  return result;
-}
-
-void RemoteObject::maybeInitializeOnUIRuntime(jsi::Runtime &rt) {
-  if (initializer.get() != nullptr) {
-    backing = initializer->shallowClone(rt);
-    initializer = nullptr;
-  }
-}
-
-jsi::Value RemoteObject::get(jsi::Runtime &rt, const jsi::PropNameID &name) {
-  if (module->isUIRuntime(rt)) {
-    return backing->getProperty(rt, name);
-  }
-  return jsi::Value::undefined();
-}
-
-void RemoteObject::set(jsi::Runtime &rt, const jsi::PropNameID &name, const jsi::Value &value) {
-  if (module->isUIRuntime(rt)) {
-    backing->setProperty(rt, name, value);
-  }
-  // TODO: we should throw if trying to update remote from host runtime
-}
-
-std::vector<jsi::PropNameID> RemoteObject::getPropertyNames(jsi::Runtime &rt) {
-  std::vector<jsi::PropNameID> res;
-  auto propertyNames = backing->getPropertyNames(rt);
-  for (size_t i = 0, size = propertyNames.size(rt); i < size; i++) {
-    res.push_back(jsi::PropNameID::forString(rt, propertyNames.getValueAtIndex(rt, i).asString(rt)));
-  }
-  return res;
 }
 
 }
