+++ Common/cpp/headers/SharedItems/MutableValueSetterProxy.h
@@ -1,0 +1,19 @@
+#pragma once
+
+#include "SharedParent.h"
+
+namespace reanimated {
+
+using namespace facebook;
+
+class MutableValueSetterProxy: public jsi::HostObject {
+private:
+  friend MutableValue;
+  std::shared_ptr<MutableValue> mutableValue;
+public:
+  MutableValueSetterProxy(std::shared_ptr<MutableValue> mutableValue): mutableValue(std::move(mutableValue)) {}
+  void set(jsi::Runtime &rt, const jsi::PropNameID &name, const jsi::Value &value);
+  jsi::Value get(jsi::Runtime &rt, const jsi::PropNameID &name);
+};
+
+}
